+++ src/items/armor.c
@@ -102,7 +102,7 @@ void load_armor_effects(world_t *world) {
 			}
 			int idx = item_data[i].stat_type.armor.modifier_count;
 			
-			item_data[i].stat_type.armor.modifier_stats[idx].modifier = atoi(value);
+			item_data[i].stat_type.armor.modifier_stats[idx].modifier = (int16_t)atoi(value);
 			item_data[i].stat_type.armor.modifier_stats[idx].stat = get_stat(target_stat);
 			item_data[i].stat_type.armor.equipped = false;
 			item_data[i].stat_type.armor.modifier_count++;
@@ -148,7 +148,7 @@ void load_armor_data(world_t *world) {
 					item_data[world->item_data_count].stat_type.armor.type = armor_get_type(token);
 					break;
 				case 2:
-					item_data[world->item_data_count].stat_type.armor.defense = atoi(token);
+					item_data[world->item_data_count].stat_type.armor.defense = (int16_t)atoi(token);
 					break;
 			}
 			token = strtok(NULL, ",");
